CDOM: Code changes for Unit Tests + Additional Test
git-svn-id: svn+ssh://svn.code.sf.net/p/pcgen/code/branches/cdom@2685 ef53c05b-3124-4d59-ae7d-fe3273f42be7

+++ code/src/java/pcgen/cdom/enumeration/ObjectKey.java
@@ -127,7 +127,7 @@ public final class ObjectKey<T> {
 
 	public static final ObjectKey<Boolean> EXCLUSIVE = new ObjectKey<Boolean>();
 
-	public static final ObjectKey<Double> CONTAINER_WEIGHT_CAPACITY = new ObjectKey<Double>();
+	public static final ObjectKey<BigDecimal> CONTAINER_WEIGHT_CAPACITY = new ObjectKey<BigDecimal>();
 
 	public static final ObjectKey<Boolean> CONTAINER_CONSTANT_WEIGHT = new ObjectKey<Boolean>();
